# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AltaPersonaRequest" : {
        "properties" : {
          "documentoIdentidad" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "fechaAlta" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "nombre" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "puesto" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "rol" : {
            "enum" : [ "ADMIN", "ENCARGADO", "EMPLEADO", "REPRESENTANTE" ],
            "type" : [ "string", "null" ]
          },
          "tipoContrato" : {
            "minLength" : 1,
            "pattern" : "JORNADA_COMPLETA|PARCIAL|POR_HORAS",
            "type" : "string"
          }
        },
        "required" : [ "documentoIdentidad", "email", "fechaAlta", "nombre", "puesto", "rol", "tipoContrato" ],
        "type" : "object"
      },
      "AltaPersonaResponse" : {
        "properties" : {
          "cuentaId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "empleadoId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "fichaCreada" : {
            "type" : "boolean"
          },
          "passwordTemporal" : {
            "type" : "string"
          },
          "rol" : {
            "enum" : [ "ADMIN", "ENCARGADO", "EMPLEADO", "REPRESENTANTE" ],
            "type" : "string"
          }
        },
        "required" : [ "cuentaId", "email", "empleadoId", "fichaCreada", "passwordTemporal", "rol" ],
        "type" : "object"
      },
      "AusenciaResponse" : {
        "properties" : {
          "canceladaEn" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "causa" : {
            "enum" : [ "MATRIMONIO", "NACIMIENTO", "FALLECIMIENTO_FAMILIAR", "ENFERMEDAD_FAMILIAR", "MUDANZA", "DEBER_INEXCUSABLE", "OTRO" ],
            "type" : [ "string", "null" ]
          },
          "comentario" : {
            "type" : [ "string", "null" ]
          },
          "desde" : {
            "format" : "date",
            "type" : "string"
          },
          "empleadoId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "estado" : {
            "enum" : [ "PENDIENTE", "APROBADA", "RECHAZADA", "CANCELADA" ],
            "type" : "string"
          },
          "hasta" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "motivoRechazo" : {
            "type" : [ "string", "null" ]
          },
          "resueltaEn" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "resueltaPor" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "solicitadaEn" : {
            "format" : "date-time",
            "type" : "string"
          },
          "solicitadaPor" : {
            "format" : "uuid",
            "type" : "string"
          },
          "tipo" : {
            "enum" : [ "VACACIONES", "PERMISO", "BAJA_MEDICA" ],
            "type" : "string"
          }
        },
        "required" : [ "desde", "empleadoId", "estado", "id", "solicitadaEn", "solicitadaPor", "tipo" ],
        "type" : "object"
      },
...
      "Error" : {
        "description" : "Formato único de error (principio VIII)",
        "properties" : {
          "code" : {
            "description" : "Identificador estable, para programas",
            "example" : "VALIDACION",
            "type" : "string"
          },
          "message" : {
            "description" : "Texto para personas; puede cambiar",
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ],
        "type" : "object"
      },
...
      "RegistroAusenciaRequest" : {
        "properties" : {
          "causa" : {
            "enum" : [ "MATRIMONIO", "NACIMIENTO", "FALLECIMIENTO_FAMILIAR", "ENFERMEDAD_FAMILIAR", "MUDANZA", "DEBER_INEXCUSABLE", "OTRO" ],
            "type" : [ "string", "null" ]
          },
          "comentario" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "desde" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "empleadoId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "hasta" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "tipo" : {
            "enum" : [ "VACACIONES", "PERMISO", "BAJA_MEDICA" ],
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "desde", "empleadoId", "tipo" ],
        "type" : "object"
      },
      "RegistroRequest" : {
        "properties" : {
          "codigoArranque" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "documentoIdentidad" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "nombre" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "codigoArranque", "documentoIdentidad", "email", "nombre", "password" ],
        "type" : "object"
      },
      "RegistroResponse" : {
        "properties" : {
          "estado" : {
            "type" : "string"
          },
          "mensaje" : {
            "type" : "string"
          }
        },
        "required" : [ "estado", "mensaje" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "API de Granatum: inventario, registro de jornada, acceso y registro, facturación, ausencias y notificaciones. Autenticación con token Bearer (JWT) obtenido en POST /api/auth/login. Los errores siguen siempre el formato {code, message}; los códigos de cada ruta están en specs/<feature>/contracts/README.md.",
    "title" : "Granatum Suite API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/ausencias/registro" : {
      "post" : {
        "operationId" : "registrar_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegistroAusenciaRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AusenciaResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Petición no válida (VALIDACION)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Sin token, o token caducado (NO_AUTENTICADO, TOKEN_ACCESO_EXPIRADO)"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "El rol no permite esta operación (FORBIDDEN)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Demasiadas peticiones desde esta dirección (DEMASIADAS_PETICIONES); ver Retry-After"
          }
        },
        "tags" : [ "ausencia-controller" ]
      }
    },
...
    "/api/auth/altas" : {
      "post" : {
        "operationId" : "darDeAlta",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AltaPersonaRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AltaPersonaResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Petición no válida (VALIDACION)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Sin token, o token caducado (NO_AUTENTICADO, TOKEN_ACCESO_EXPIRADO)"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "El rol no permite esta operación (FORBIDDEN)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Demasiadas peticiones desde esta dirección (DEMASIADAS_PETICIONES); ver Retry-After"
          }
        },
        "tags" : [ "alta-persona-controller" ]
      }
    },
...
    "/api/auth/registro" : {
      "post" : {
        "operationId" : "registrar",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegistroRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RegistroResponse"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Petición no válida (VALIDACION)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Demasiadas peticiones desde esta dirección (DEMASIADAS_PETICIONES); ver Retry-After"
          }
        },
        "security" : [ ],
        "tags" : [ "registro-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}